# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion de Gestion Promos\nPOUR UNE FACILITATION DES TESTS, L'AUTHENTIFICATION EST TEMPORAIREMENT DÉSACTIVÉE !",
        "version": "1.0.0",
        "description": "API Gestion-Promos"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": ["users"],
                "summary": "Lister tous les utilisateurs",
                "description": "Recupére la liste de tous les utilisateurs",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": ["users"],
                "summary": "Créer un nouvel utilisateur",
                "description": "Endpoint pour créer un nouvel utilisateur avec les données spécifiées.",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "pattern": "^((77|76|75|70|78)\\d{3}\\d{2}\\d{2})|(33[8]\\d{2}\\d{2}\\d{2})$"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "Admin",
                                            "CM",
                                            "Manager",
                                            "Coach",
                                            "Apprenant"
                                        ]
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Photo de profil à télécharger",
                                        "maxLength": 1024
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": ["actif", "bloqué"]
                                    }
                                },
                                "required": [
                                    "nom",
                                    "prenom",
                                    "adresse",
                                    "telephone",
                                    "role",
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nom": {
                                            "type": "string"
                                        },
                                        "prenom": {
                                            "type": "string"
                                        },
                                        "adresse": {
                                            "type": "string"
                                        },
                                        "telephone": {
                                            "type": "string"
                                        },
                                        "role": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "photo": {
                                            "type": "string",
                                            "format": "binary"
                                        },
                                        "statut": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}